+++ src/lib.rs
@@ -1,7 +1,7 @@
 /*
 MIT License
 
-Copyright (c) 2021 Philipp Schuster
+Copyright (c) 2024 Philipp Schuster
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
@@ -65,6 +65,7 @@ use std::io::{Read as IoRead, Write as IoWrite};
 use std::net::{IpAddr, TcpStream};
 use std::str::FromStr;
 use std::sync::Arc;
+use std::thread;
 use std::time::{Duration, Instant};
 use trust_dns_resolver::Resolver as DnsResolver;
 use url::Url;
@@ -167,7 +168,7 @@ fn tls_handshake_if_necessary(
                 .into()
         } else {
             RustlsConnector::new_with_native_certs()
-                .map_or_else(|_| RustlsConnector::new_with_webpki_roots_certs(), |v| v)
+                .unwrap_or_else(|_| RustlsConnector::new_with_webpki_roots_certs())
         };
         let now = Instant::now();
 
@@ -274,13 +275,17 @@ fn execute_http_get(
 }
 
 /// Constructs the header for a HTTP/1.1 GET-Request.
+///
+/// Sets the following default headers:
+/// - `Accept-Encoding: gzip, deflate, br, zstd` (default of Chrome v123)
+/// - `User-Agent: ttfb/<version>`
 fn build_http11_header(url: &Url) -> String {
     format!(
         "GET {path} HTTP/1.1\r\n\
         Host: {host}\r\n\
         User-Agent: ttfb/{version}\r\n\
         Accept: */*\r\n\
-        Accept-Encoding: gzip, deflate, br\r\n\
+        Accept-Encoding: gzip, deflate, br, zstd\r\n\
         \r\n",
         path = url.path(),
         host = url.host_str().unwrap(),
@@ -322,18 +327,31 @@ fn check_scheme_is_allowed(url: &Url) -> Result<(), TtfbError> {
 /// If the user gave us a domain name, we resolve it using the [`trust-dns-resolver`]
 /// crate and measure the time for it.
 fn resolve_dns_if_necessary(url: &Url) -> Result<(IpAddr, Option<Duration>), TtfbError> {
-    Ok(if url.domain().is_none() {
-        let mut ip_str = url.host_str().unwrap();
-        // [a::b::c::d::e::f::0::1] => ipv6 address
-        if ip_str.starts_with('[') {
-            ip_str = &ip_str[1..ip_str.len() - 1];
+    match url.domain() {
+        Some(domain) => {
+            // shortcut
+            if domain.eq("localhost") {
+                Ok((
+                    IpAddr::from_str("127.0.0.1").unwrap(),
+                    Some(Duration::default()),
+                ))
+            } else {
+                resolve_dns(url).map(|(addr, dur)| (addr, Some(dur)))
+            }
         }
-        let addr = IpAddr::from_str(ip_str)
-            .map_err(|e| TtfbError::InvalidUrl(InvalidUrlError::WrongFormat(e.to_string())))?;
-        (addr, None)
-    } else {
-        resolve_dns(url).map(|(addr, dur)| (addr, Some(dur)))?
-    })
+        None => {
+            let mut ip_str = url.host_str().unwrap();
+            // [a::b::c::d::e::f::0::1] => ipv6 address
+            let is_ipv6_addr = ip_str.starts_with('[');
+            if is_ipv6_addr {
+                ip_str = &ip_str[1..ip_str.len() - 1];
+            }
+            let addr = IpAddr::from_str(ip_str)
+                .map_err(|e| TtfbError::InvalidUrl(InvalidUrlError::WrongFormat(e.to_string())))?;
+
+            Ok((addr, None))
+        }
+    }
 }
 
 /// Actually resolves a domain using the systems default DNS resolver.
@@ -349,11 +367,28 @@ fn resolve_dns(url: &Url) -> Result<(IpAddr, Duration), TtfbError> {
 
     let begin = Instant::now();
 
-    // at least on Linux this gets cached somehow in the background
-    // probably the DNS implementation/OS has a DNS cache
-    let response = resolver
-        .lookup_ip(url.host_str().unwrap())
-        .map_err(|err| TtfbError::CantResolveDns(ResolveDnsError::Other(Box::new(err))))?;
+    // At least on Linux this gets cached somehow in the background
+    // probably the DNS implementation/OS has a DNS cache.
+    //
+    // Internally, the resolver library spawns a tokio runtime. The only way to
+    // prevent "cannot start a runtime from within a runtime" from tokio is to
+    // move this invocation to a dedicated thread. While this is more a
+    // workaround than a nice solution, it is simple and effective.
+    //
+    // For the performance/measurements, this change is negligible.
+    //
+    // More info: https://stackoverflow.com/a/62536772/2891595
+    let response = {
+        let url = url.clone();
+        thread::spawn(move || {
+            resolver
+                .lookup_ip(url.host_str().unwrap())
+                .map_err(|err| TtfbError::CantResolveDns(ResolveDnsError::Other(Box::new(err))))
+        })
+        .join()
+        .unwrap()?
+    };
+
     let duration = begin.elapsed();
 
     let ipv4_addrs = response
@@ -376,8 +411,6 @@ fn resolve_dns(url: &Url) -> Result<(IpAddr, Duration), TtfbError> {
 
 #[cfg(all(test, not(network_tests)))]
 mod tests {
-    use crate::parse_input_as_url;
-
     use super::*;
 
     #[test]
@@ -392,6 +425,8 @@ mod tests {
         parse_input_as_url("https://goo-gle.com:443/foobar?124141").expect("to be valid");
         parse_input_as_url("https://subdomain.goo-gle.com:443/foobar?124141").expect("to be valid");
         parse_input_as_url("https://192.168.1.102:443/foobar?124141").expect("to be valid");
+
+        parse_input_as_url("http://localhost").expect("to be valid");
     }
 
     #[test]
@@ -417,6 +452,18 @@ mod tests {
         assert_eq!(
             prepend_default_scheme_if_necessary("ftp://192.168.1.102:443/foobar?124141".to_owned()),
             "ftp://192.168.1.102:443/foobar?124141"
+        );
+    }
+
+    #[test]
+    fn test_dns_if_necessary_localhost_shortcut() {
+        let url = url::Url::from_str("http://localhost").unwrap();
+        assert_eq!(
+            resolve_dns_if_necessary(&url),
+            Ok((
+                IpAddr::from_str("127.0.0.1").unwrap(),
+                Some(Duration::from_secs(0))
+            ))
         );
     }